+++ backend/src/main/java/com/carpentry/manager/dashboard/dto/DashboardStats.java
@@ -4,11 +4,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 public class DashboardStats {
     private long totalItems;
     private long openOrders; // Logic: Items with low stock (e.g., < 5)
-    private double totalCustomerDebt;
+    private BigDecimal totalCustomerDebt;
 }
