+++ backend/src/main/java/com/carpentry/manager/report/dto/MonthlyFinancialReport.java
@@ -4,13 +4,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 public class MonthlyFinancialReport {
     private int year;
     private int month;
-    private double totalIncome;
-    private double totalExpenses;
-    private double netProfit;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpenses;
+    private BigDecimal netProfit;
 }
